refactor: optimisation de la page level
